% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{cauchy.test}
\alias{cauchy.test}
\title{Function for testing if a sample is cauchy distributed.}
\usage{
cauchy.test(Test, Sample, parameter = NULL, method_estimation = 2,
  alpha = 0.05, repetitions = 10000, data_print = TRUE)
}
\arguments{
\item{Test}{A string with the name of the test.}

\item{Sample}{A numeric vector with data.}

\item{parameter}{A positive numeric value if the test requires a parameter (see below).}

\item{method_estimation}{A parameter specifying the method for estimating the Cauchy-distribution parameters.}

\item{alpha}{A numeric value between 0 and 1 specifying the significance level.}

\item{repetitions}{A positiv integer

Possible tests are D_Henze, KL, W, AD, CM, KS, T1, T2, T3, T4. The tests D_Henze, T1, T2, T3, T4 require a positive parameter, which influences the weight-function.}
}
\description{
\code{cauchy.test} returns if a sample is cauchy distributed with a given significance level.
}
