% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{D}
\alias{D}
\title{D - Teststatistic}
\usage{
D(y, l)
}
\arguments{
\item{y}{A numeric vector.}

\item{l}{A positive number.}
}
\value{
A number.
}
\description{
\code{D} returns the value gained by the D test
}
\details{
This function computes the value of the D Test with weighth parameter l.
}
\seealso{
Other teststatistic: \code{\link{T1}}
}
\concept{teststatistic}
