% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{distr}
\alias{distr}
\title{Function for standardizing data}
\usage{
distr(n, case)
}
\arguments{
\item{n}{The size of the dataset (10,20 or 50).}

\item{case}{A name of a distribution.}
}
\description{
distr returns a random vector simulated by a choosen distribution.
}
