% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{median_est}
\alias{median_est}
\alias{ml_est}
\alias{eise_est}
\title{Parameter estimation}
\usage{
median_est(x)

ml_est(x)

eise_est(x, l)
}
\arguments{
\item{x}{A Numeric vector.}

\item{l}{Parameter for the weight function described in \code{\link{D}}}
}
\value{
the location and the scale parameter.
}
\description{
\code{median_est} returns the estimated parameters of a cauchy distribution.
}
\details{
This function ....
}
\section{Functions}{
\itemize{
\item \code{ml_est}: Parameter estimation with the ML Method

\item \code{eise_est}: Paremter estimation with an estimated squared integrated error.
}}

