% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{quantile_edf_Median}
\alias{quantile_edf_Median}
\title{Quantile of the edf statistics with Median parameter estimation.}
\format{A data frame with 4 rows and 1 columns.}
\usage{
quantile_edf_Median
}
\description{
A dataset contatining the 0.95 quantile of the D statistic gained by a 100k rep Monte-Carlo simulation.
Each row is for one of the classic test (KS, CM, AD,W) and the row is for the datasize (10,20,50,100,200)
}
\keyword{datasets}
