% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{quantile_edf_ML}
\alias{quantile_edf_ML}
\title{Quantile of the edf statistics with ML parameter estimation.}
\format{A data frame with 5 rows and 3 columns.}
\usage{
quantile_edf_ML
}
\description{
A dataset containing the 0.95 quantile of the D statistic gained by a 100k rep Monte-Carlo simulation.
Each row is for one of the classic test (KS, CM, AD,W) and the row is for the datasize (10,20,50,100,200)
}
\keyword{datasets}
