% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{quantile_KL_Median}
\alias{quantile_KL_Median}
\title{Quantile of the KL statistic with Median parameter estimation.}
\format{A data frame with 4 rows and 1 columns.}
\usage{
quantile_KL_Median
}
\description{
A dataset contatining the 0.95 quantile of the D statistic gained by a 100k rep Monte-Carlo simulation.
}
\keyword{datasets}
