% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{quantile_KL_ML}
\alias{quantile_KL_ML}
\title{Quantile of the KL statistic with ML parameter estimation.}
\format{A data frame with 4 rows and 1 columns.}
\usage{
quantile_KL_ML
}
\description{
A dataset containing the 0.95 quantile of the D statistic gained by a 100k rep Monte-Carlo simulation.
}
\keyword{datasets}
