% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{T1}
\alias{T1}
\alias{T2}
\alias{T3}
\alias{T4}
\alias{T5}
\title{T1 - Teststatistic}
\usage{
T1(y, l = 1)

T2(y, l = 1)

T3(y, l = 1)

T4(y, l = 1)

T5(y, l = 1)
}
\arguments{
\item{y}{A numeric vector.}

\item{l}{A positive number.}
}
\value{
A number.
}
\description{
The function returns the value gained by the
\enumerate{
     \item T1 Test,
     \item T2 Test,
     \item T3 Test,
     \item T4 Test,
     \item T5 Test.}
}
\details{
This function computes the value of the T Test with weighth parameter l. For T5 l cannot be 1!
}
\section{Functions}{
\itemize{
\item \code{T2}: T2 - Statistic

\item \code{T3}: T3 - Statistic

\item \code{T4}: T4 - Statistic

\item \code{T5}: T5 - Statistic
}}

\seealso{
Other teststatistic: \code{\link{D}}
}
\concept{teststatistic}
