% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{testentscheid}
\alias{testentscheid}
\title{Function for the result of a test}
\usage{
testentscheid(n, Verteilung = "", method, x = 0)
}
\arguments{
\item{n}{An integer. (Here 10,20 or 50)}

\item{Verteilung}{A String specifing the choosen distribution to test.}

\item{method}{A name of a test \code{\link{teststatistic}}}

\item{x}{A numeric vector of length n.}
}
\value{
binary number.
}
\description{
\code{testentscheid} returns 0 if \eqn{H_0} is not rejected and 1 if it is rejected.
}
\details{
If \code{Verteilung} is empty then the given numeric vector x is taken for testing.
}
