% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{tfc}
\alias{tfc}
\alias{tfc-package}
\title{tfc: A package for testing for Cauchy-distribution}
\description{
The tfc package provides multiple test and 3 ways to estimate paramteres of a Cauchy-Distribution.
}
\section{tfc functions}{

\itemize{
\item \code{\link{D}}
\item \code{\link{distr}}
\item \code{\link{eise_est}}
\item \code{\link{median_est}}
\item \code{\link{ml_est}}
\item \code{\link{standardisiert}}
\item \code{\link{T1}}
\item \code{\link{T2}}
\item \code{\link{T3}}
\item \code{\link{T4}}
\item \code{\link{T5}}
\item \code{\link{testentscheid}}
}
}

\section{tfc data}{

\itemize{
\item \code{\link{quantile_D_Median}}
\item \code{\link{quantile_D_ML}}
}
}

