% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{standardisiert}
\alias{standardisiert}
\title{Function for standardizing data}
\usage{
standardisiert(x, method_estimation = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{method_estimation}{A Parameter.}
}
\value{
A numeric vector.
}
\description{
\code{standardisiert} returns the standarized vector.
}
\details{
This function standardizes the Inputvector. This is done with a parameter estimation function like \code{\link{median_est}}.
For  \enumerate{
             \item \eqn{method_estimation=1} \code{\link{median_est}} is used,
             \item \eqn{method_estimation=2} \code{\link{ml_est}} is used,
             \item \eqn{method_estimation=3} \code{\link{eise_est}} is used.
}
}
